refactor: 테스트 일관성을 위한 sql script 파일 사용에 따른 연관 코드 수정
